cleanup the exposition of status data and ensure the module registry provides enough info
