+++ kamon-core/src/main/resources/reference.conf
@@ -44,6 +44,8 @@ kamon {
   # kamon.modules {
   #   module-name {
   #     enabled = true
+  #     description = "A module description"
+  #     kind = "combined | metric | span | plain"
   #     class = "com.example.ModuleClass"
   #   }
   # }
