+++ kamon-core/src/main/scala/kamon/metric/MetricRegistry.scala
@@ -26,6 +26,7 @@ import scala.collection.concurrent.TrieMap
 import java.time.Duration
 import java.util.concurrent.ScheduledExecutorService
 
+import kamon.status.Status
 import org.slf4j.LoggerFactory
 
 
@@ -89,22 +90,17 @@ class MetricRegistry(initialConfig: Config, scheduler: ScheduledExecutorService)
     metric.asInstanceOf[T]
   }
 
-  private[kamon] def status(): MetricRegistry.Status = {
-    var metricInfos = Seq.empty[MetricRegistry.MetricInfo]
+  private[kamon] def status(): Status.MetricRegistry = {
+    var registeredMetrics = Seq.empty[Status.Metric]
     metrics.foreach {
       case (metricName, metric) =>
         metric.incarnations().foreach(incarnation => {
-          metricInfos = metricInfos :+ MetricRegistry.MetricInfo(metricName, incarnation, metric.instrumentType)
+          registeredMetrics = registeredMetrics :+ Status.Metric(metricName, incarnation, metric.instrumentType)
         })
     }
 
-    MetricRegistry.Status(metricInfos)
+    Status.MetricRegistry(registeredMetrics)
   }
-}
-
-object MetricRegistry {
-  case class Status(metrics: Seq[MetricInfo])
-  case class MetricInfo(name: String, tags: Map[String, String], instrumentType: InstrumentType)
 }
 
 trait MetricsSnapshotGenerator {
